requirements.txt
beautifulsoup4
free-proxy
prettytable
python-mpv
git+https://github.com/v1k45/python-qBittorrent
requests
retry
PyQt5
shikimori_api
anitopy
